fix size_t ocurrences

+++ deliver/ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: gaguiar- <gaguiar-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/22 18:32:34 by gaguiar-          #+#    #+#             */
-/*   Updated: 2025/07/23 11:21:20 by gaguiar-         ###   ########.fr       */
+/*   Updated: 2025/07/23 11:41:05 by gaguiar-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 void	*ft_memset(void *s, int c, size_t n)
 {
-	int				i;
-	unsigned char	*aux;
+	size_t				i;
+	unsigned char		*aux;
 
 	aux = (unsigned char *)s;
 	i = 0;
