+++ deliver/ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: gaguiar- <gaguiar-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/21 11:51:23 by gaguiar-          #+#    #+#             */
-/*   Updated: 2025/07/23 11:24:03 by gaguiar-         ###   ########.fr       */
+/*   Updated: 2025/07/23 11:35:57 by gaguiar-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_strlcpy(char *dst, const char *src, size_t size)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	if (size > 0)
