+++ deliver/ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: gaguiar- <gaguiar-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/21 16:01:33 by gaguiar-          #+#    #+#             */
-/*   Updated: 2025/07/23 11:23:30 by gaguiar-         ###   ########.fr       */
+/*   Updated: 2025/07/23 11:39:24 by gaguiar-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strnstr(const char *big, const char *little, size_t len)
 {
-	int		i;
-	int		j;
+	size_t		i;
+	size_t		j;
 
 	i = 0;
 	if (!*little)
