+++ deliver/libft.h
@@ -6,7 +6,7 @@
 /*   By: gaguiar- <gaguiar-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/19 14:05:33 by gaguiar-          #+#    #+#             */
-/*   Updated: 2025/07/23 11:20:23 by gaguiar-         ###   ########.fr       */
+/*   Updated: 2025/07/23 11:33:18 by gaguiar-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,14 +20,14 @@ int		ft_isalpha(int c);
 int		ft_isascii(int c);
 int		ft_isdigit(int c);
 int		ft_isprint(int c);
-void	*ft_memset(void *s, int c, int n);
-int		ft_strlcat(char *dst, const char *src, int size);
-int		ft_strlcpy(char *dst, const char *src, int size);
+void	*ft_memset(void *s, int c, size_t n);
+int		ft_strlcat(char *dst, const char *src, size_t size);
+int		ft_strlcpy(char *dst, const char *src, size_t size);
 int		ft_strlen(const char *s);
-int		ft_strncmp(const char *s1, const char *s2, int n);
-char	*ft_strnstr(const char *big, const char *little, int len);
+int		ft_strncmp(const char *s1, const char *s2, size_t n);
+char	*ft_strnstr(const char *big, const char *little, size_t len);
 int		ft_tolower(int c);
 int		ft_toupper(int c);
-void	ft_bzero(void *s, int n);
+void	ft_bzero(void *s, size_t n);
 
 #endif
